requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
Django==5.1.7
django-multiselectfield==0.1.13
djangorestframework==3.15.2
exceptiongroup==1.2.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
yarl==1.18.3
